Process assay tab data

Process data from assay tab file

Process data from assay tab files with technology type mass spectrometry (ms). Processing those files requires the xcms package to be installed.

Process data from assay tab files with technology type DNA microarray (ms). Processing those files requires the Biobase and affy packages to be installed.

Usage

processAssay(isaObject, aTabObject, type = c("raw", "derived"))

# S4 method for class 'ISAjson,assayTab,character'
processAssay(isaObject, aTabObject, type = c("raw", "derived"))

# S4 method for class 'ISAjson,msAssayTab,character'
processAssay(isaObject, aTabObject, type = c("raw", "derived"))

# S4 method for class 'ISAjson,microarrayAssayTab,character'
processAssay(isaObject, aTabObject, type = c("raw", "derived"))

Arguments

isaObject

An object of the ISAjson.

aTabObject

An object of the microarrayAssayTab.

type

A character string indicating which data files should be processed, either "raw" for raw data files, or "derived" for derived data files. The file names are taken from the corresponding column in the aTabObject.