Read an ISA-JSON file into an R object.

Reads an ISA-JSON file, given as a zip file or as a json file in a specific folder, and builds an object of the ISAjson class.

Usage

readISAJSON(file, zipfile = NULL, verbose = FALSE)

Arguments

file

A character string with the name of the ISA-JSON file is located (if the parameter zipfile is not provided or if it is equal to NULL), or the name of the directory where the zip archive containing the ISA-JSON file is located (if the parameter zipfile is not NULL).

zipfile

A character vector with the name of the zip archive containing ISA-JSON file itself (without a directory name in the zip archive). The default value is NULL (specifying that the ISA-JSON file has not been archived in a zipped file).

verbose

Should the messages for the different reading steps be shown?

Value

An object of the ISAjson class.