Check the validity of an object of class ISAjson.

The validISAObject function checks whether an object of class ISAjson is a valid object. An object of the ISAjson is considered valid when: * The structure of the json file matches the schema for a valid ISA-JSON file that can be found at https://github.com/ISA-tools/isa-api/tree/master/isatools/resources/schemas/isa_model_version_1_0_schemas/core *

Usage

validISAJSONObject(object)

Arguments

object

An object of class ISAjson.

Value

TRUE or an error message.

See also